move gain caches to FM folder
